requirements.txt
streamlit
python-docx
requests
anthropic
PyPDF2
wikipediaapi